register updater, dividers, and serializers in the module __init__.py
